add gui integration
