+++ src/main/java/gui/DialogBox.java
@@ -1,3 +1,5 @@
+package gui;
+
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.geometry.Pos;
